injury_predict_web/requirements.txt
Django==2.2.7
djangorestframework==3.10.3
mysqlclient==1.4.5
numpy==1.21.2
pandas==1.3.3
patsy==0.5.2
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.1
pyxlsb==1.0.5
scipy==1.7.1
six==1.12.0
SQLAlchemy==1.3.10
sqlparse==0.3.0
xlrd==1.2.0
django-cryptographic-fields
requests
django-background-tasks==1.2.5
cython
convertdate
LunarCalendar
holidays
matplotlib
plotly
statsmodels==0.13.0
openpyxl